Move networking/analytics packages from collect to root package
