+++ vgscheckout/src/main/java/com/verygoodsecurity/vgscheckout/analytic/AnalyticTracker.kt
@@ -1,0 +1,10 @@
+package com.verygoodsecurity.vgscheckout.analytic
+
+import com.verygoodsecurity.vgscheckout.analytic.event.core.Event
+
+internal interface AnalyticTracker {
+
+    var isEnabled: Boolean
+
+    fun log(event: Event)
+}
