+++ vgscheckout/src/main/java/com/verygoodsecurity/vgscheckout/analytic/event/CancelEvent.kt
@@ -1,0 +1,10 @@
+package com.verygoodsecurity.vgscheckout.analytic.event
+
+import com.verygoodsecurity.vgscheckout.analytic.event.core.Event
+
+private const val TYPE = "Cancel"
+
+internal object CancelEvent : Event(TYPE) {
+
+    override val attributes: Map<String, Any> = emptyMap()
+}
