--- vgscheckout/src/main/java/com/verygoodsecurity/vgscheckout/collect/core/analytic/event/ScanEvent.kt
@@ -1,24 +1,0 @@
-package com.verygoodsecurity.vgscheckout.collect.core.analytic.event
-
-import com.verygoodsecurity.vgscheckout.collect.core.analytic.event.core.Event
-
-internal data class ScanEvent(
-    val status: String,
-    val scannerType: String,
-    val scanId: String?
-) : Event(TYPE) {
-
-    override val attributes: Map<String, Any> = LinkedHashMap<String, Any>().apply {
-        put(KEY_STATUS, status)
-        put(KEY_SCANNER_TYPE, scannerType)
-        scanId?.let { put(KEY_SCAN_ID, scanId) }
-    }
-
-    companion object {
-
-        private const val TYPE = "Scan"
-
-        private const val KEY_SCANNER_TYPE = "scannerType"
-        private const val KEY_SCAN_ID = "scanId"
-    }
-}
