--- vgscheckout/src/main/java/com/verygoodsecurity/vgscheckout/collect/core/networking/client/HttpBodyFormat.kt
@@ -1,6 +1,0 @@
-package com.verygoodsecurity.vgscheckout.collect.core.networking.client
-
-internal enum class HttpBodyFormat constructor(val value: String) {
-    JSON("application/json"),
-    X_WWW_FORM_URLENCODED("application/x-www-form-urlencoded")
-}
