+++ vgscheckout/src/main/java/com/verygoodsecurity/vgscheckout/networking/client/HttpBodyFormat.kt
@@ -1,0 +1,6 @@
+package com.verygoodsecurity.vgscheckout.networking.client
+
+internal enum class HttpBodyFormat constructor(val value: String) {
+    JSON("application/json"),
+    X_WWW_FORM_URLENCODED("application/x-www-form-urlencoded")
+}
