+++ vgscheckout/src/main/java/com/verygoodsecurity/vgscheckout/networking/client/HttpMethod.kt
@@ -1,0 +1,34 @@
+package com.verygoodsecurity.vgscheckout.networking.client
+
+/**
+ *  The set of common methods for HTTP.
+ *
+ *  @version 1.0.0
+ */
+internal enum class HttpMethod {
+
+    /**
+     *  HTTP GET method
+     */
+    GET,
+
+    /**
+     *  HTTP POST method
+     */
+    POST,
+
+    /**
+     *  HTTP PUT method
+     */
+    PUT,
+
+    /**
+     *  HTTP PATCH method
+     */
+    PATCH,
+
+    /**
+     *  HTTP DELETE method
+     */
+    DELETE
+}
