+++ vgscheckout/src/main/java/com/verygoodsecurity/vgscheckout/ui/core/BaseCheckoutActivity.kt
@@ -5,7 +5,7 @@ import android.os.Bundle
 import androidx.annotation.CallSuper
 import androidx.appcompat.app.AppCompatActivity
 import com.verygoodsecurity.vgscheckout.R
-import com.verygoodsecurity.vgscheckout.collect.core.analytic.event.CancelEvent
+import com.verygoodsecurity.vgscheckout.analytic.event.CancelEvent
 import com.verygoodsecurity.vgscheckout.config.core.CheckoutConfig
 import com.verygoodsecurity.vgscheckout.model.CheckoutResultContract
 import com.verygoodsecurity.vgscheckout.ui.fragment.core.BaseFragment
