+++ vgscheckout/src/main/java/com/verygoodsecurity/vgscheckout/ui/fragment/save/SaveCardFragment.kt
@@ -12,8 +12,8 @@ import androidx.annotation.VisibleForTesting
 import androidx.core.view.isGone
 import androidx.core.view.isVisible
 import com.verygoodsecurity.vgscheckout.R
-import com.verygoodsecurity.vgscheckout.collect.core.analytic.event.RequestEvent
-import com.verygoodsecurity.vgscheckout.collect.core.analytic.event.ResponseEvent
+import com.verygoodsecurity.vgscheckout.analytic.event.RequestEvent
+import com.verygoodsecurity.vgscheckout.analytic.event.ResponseEvent
 import com.verygoodsecurity.vgscheckout.collect.core.model.network.VGSError
 import com.verygoodsecurity.vgscheckout.collect.core.storage.InternalStorage
 import com.verygoodsecurity.vgscheckout.collect.view.InputFieldView
