+++ vgscheckout/src/main/java/com/verygoodsecurity/vgscheckout/util/extension/CheckoutConfig.kt
@@ -2,7 +2,7 @@ package com.verygoodsecurity.vgscheckout.util.extension
 
 import android.content.Context
 import com.verygoodsecurity.vgscheckout.R
-import com.verygoodsecurity.vgscheckout.collect.core.networking.*
+import com.verygoodsecurity.vgscheckout.networking.*
 import com.verygoodsecurity.vgscheckout.collect.core.isSandbox
 import com.verygoodsecurity.vgscheckout.config.core.CheckoutConfig
 import com.verygoodsecurity.vgscheckout.config.networking.core.getNormalizedHostName
